Automate selection of relevant build flavors for release branches (#2111)
