+++ .github/workflows/ci.yml
@@ -31,9 +31,8 @@ jobs:
         uses: actions/setup-node@v1
         with:
           node-version: '14.x'
-      # gets the matrix for ci run from the config file
       - id: get-matrix
-        run: echo "::set-output name=matrix::{\"include\":[{\"flavor\":\"stable\"},{\"flavor\":\"beta\"}]}"
+        run: echo "::set-output name=matrix::$(node ./common/scripts/workflow-read-matrix.mjs)"
 
   build_packages:
     needs: get_matrix
