--- .github/workflows/create-beta-release-branch.yml
@@ -1,115 +1,0 @@
-# Bump package version numbers and create a release branch
-name: Beta release branch - Create
-
-on:
-  # Allows you to run this workflow manually from the Actions tab
-  workflow_dispatch:
-    inputs:
-      branch:
-        description: 'Branch or tag to create release from'
-        required: true
-        default: 'main'
-
-# cancel workflow when a newer version of the workflow is triggered on the same github ref
-concurrency:
-  group: ${{ github.workflow }}-${{ github.ref }}
-  cancel-in-progress: true
-
-jobs:
-  create_beta_release:
-    name: Bump versions and create changelog PR
-    runs-on: ubuntu-latest
-    steps:
-      # Check-out repo
-      - uses: actions/checkout@v2
-        with:
-          fetch-depth: 0
-          # Use a machine account when checking out. This is to workaround the issue were GitHub
-          # actions, when using the default account, cannot trigger other actions. And we want this
-          # action to trigger the regular CI pipeline on the created branch.
-          # This machine account is only for this PAT, pwd was created and thrown away
-          # If any update needed, create a new account, add access to the repo and generate a new PAT
-          token: ${{ secrets.MACHINE_ACCOUNT_PAT }}
-
-      # Setup bot information for creating pull request
-      # Here we use the id from the github actions bot: https://api.github.com/users/better-informatics%5Bbot%5D
-      - name: Setup bot git information
-        run: |
-          git config user.email "41898282+github-actions[bot]@users.noreply.github.com"
-          git config user.name "github-actions[bot]"
-
-      # Check out onto desired branch or tag to create release from
-      - name: Checkout tag/branch
-        run: git checkout ${{ github.event.inputs.branch }}
-
-      # Ensure node version is great enough
-      - name: Use Node.js v14.x
-        uses: actions/setup-node@v1
-        with:
-          node-version: '14.x'
-
-      # Try get node_modules from cache
-      - name: Restore node_modules from cache
-        uses: actions/cache@v2
-        with:
-          path: common/temp/pnpm-store
-          key: ${{ runner.os }}-${{ hashFiles('common/config/rush/pnpm-lock.yaml') }}
-
-      # Install dependencies
-      - name: Install rush
-        run: npm install -g @microsoft/rush@5.47.0
-      - name: Install dependencies
-        run: rush install
-
-      # Bump package versions
-      - name: Bump package versions
-        run: node common/scripts/bump-beta-version.js
-
-      # Choose beta dependency for sdk
-      - name: Choose beta dependency for sdk
-        run: node common/scripts/choose-beta-sdk-deps.js
-
-      # Generate changelog using beachball
-      - name: Generate changelog using beachball
-        run: node common/scripts/beachball-changelog-only.js
-
-      - name: Synchronize package version reported to telemetry
-        run: node common/scripts/sync-telemetry-package-version
-
-      # Important to check version consistency again after bumping versions.
-      - name: Ensure all package versions are consistent
-        run: rush ensure-consistent-versions
-
-      # Retrieve new version to create branch with
-      - name: Retrieve new version from package.json
-        id: version
-        run: |
-          ver=$(jq -r .version packages/communication-react/package.json)
-          echo version: $ver
-          echo "::set-output name=version::$ver"
-
-      # Commit changes
-      - name: Hop into new branch
-        run: git checkout -b release/${{ steps.version.outputs.version }}
-      - name: Commit changes
-        run: |
-          git add .
-          git commit -m "Beta version bump"
-
-      # Create beachball change files
-      - name: Create change files
-        run: node common/config/node_modules/beachball/bin/beachball change --message 'Bump package version to ${{ steps.version.outputs.version }}' --type 'none'
-
-      # Push changes
-      - name: Push branch
-        run: git push --set-upstream origin release/${{ steps.version.outputs.version }}
-
-      # Create a PR from the changes
-      - name: Create PR
-        run: |
-          curl \
-            -X POST \
-            -H 'Accept: application/vnd.github.v3+json' \
-            -H 'authorization: Bearer ${{ secrets.GITHUB_TOKEN }}' \
-            https://api.github.com/repos/Azure/communication-ui-library/pulls \
-            -d '{ "title":"${{ steps.version.outputs.version }} release branch", "head":"release/${{ steps.version.outputs.version }}", "base":"main", "body":"Release branch for ${{ steps.version.outputs.version }}. Created by the `Beta release branch - create` GitHub action. Please review." }'
