+++ .github/workflows/create-release-branch.yml
@@ -72,14 +72,13 @@ jobs:
           git checkout -b release/${{ steps.version.outputs.version }}
           echo "::set-output name=releasebranch::release/${{ steps.version.outputs.version }}"
 
-      # Choose stable dependency for sdk
-      - name: Choose stable dependency for sdk if stable
+      - name: Force build flavor to `stable`
         if: ${{ startsWith(github.event.inputs.branch, 'prerelease-stable-minor') || startsWith(github.event.inputs.branch, 'prerelease-stable-patch')}}
-        run: node common/scripts/choose-stable-sdk-deps.js
+        run: node ./common/scripts/force-build-flavor.mjs stable
 
-      - name: Choose beta dependency for sdk if beta
+      - name: Force build flavor to `beta`
         if: ${{ startsWith(github.event.inputs.branch, 'prerelease-beta' )}}
-        run: node common/scripts/choose-beta-sdk-deps.js
+        run: node ./common/scripts/force-build-flavor.mjs beta
 
       # Synchronize the telemetry package versions.
       - name: Synchronize package version reported to telemetry
