+++ .github/workflows/update-snapshot.yml
@@ -30,7 +30,7 @@ jobs:
           node-version: '14.x'
       # gets the matrix for ci run from the config file
       - id: get-matrix
-        run: echo "::set-output name=matrix::{\"include\":[{\"flavor\":\"stable\"},{\"flavor\":\"beta\"}]}"
+        run: echo "::set-output name=matrix::$(node ./common/scripts/workflow-read-matrix.mjs)"
 
   update_examples_snapshot:
     needs: get_matrix
