+++ common/config/workflows/matrix.json
@@ -1,0 +1,11 @@
+{
+    "include": [
+        {
+            "flavor": "stable"
+        },
+        {
+            "flavor": "beta"
+        }
+
+    ]
+}
