+++ common/scripts/lib/index.mjs
@@ -1,6 +1,6 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-export * from './getBuildFlavor.mjs';
+export * from './constants.mjs';
 export * from './exec.mjs';
-
+export * from './getBuildFlavor.mjs';
